% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rddsga}
\alias{rddsga}
\title{Deprecated function name}
\usage{
rddsga(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{\link[=wsga_rdd]{wsga_rdd()}}.}
}
\value{
The return value of \code{\link[=wsga_rdd]{wsga_rdd()}}.
}
\description{
\code{rddsga()} is the previous name of \code{wsga_rdd()}. It is retained as a
deprecated alias and will be removed in a future major release. New code
should call \code{wsga_rdd()} directly.
}
\seealso{
\code{\link[=wsga_rdd]{wsga_rdd()}}
}
\keyword{internal}
